feat(helpers): deprecate methods from useNavigation hook
Deprecate setRoot from useNavigation hook
Deprecate showModal from useNavigation hook
Deprecate showOverlay from useNavigation hook
Deprecate setDefaultOptions from useNavigation hook
Deprecate dismissAllModals from useNavigation hook
Deprecate getLaunchArgs from useNavigation hook

fix #55
