+++ src/helpers/createNavigationCommands.ts
@@ -1,11 +1,9 @@
-import { Layout, LayoutRoot, Navigation, Options } from 'react-native-navigation'
+import { Layout, Navigation, Options } from 'react-native-navigation'
 import createLayout from './createLayout'
 
-interface SetRootCommand {
-  (layoutRoot: LayoutRoot): Promise<any>
-  (layout: Layout): Promise<any>
-  <P = {}>(name: string, passProps?: P, options?: Options): Promise<any>
-}
+import setRoot, { SetRootCommand } from './setRoot'
+import showModal, { ShowModalCommand } from './showModal'
+import showOverlay, { ShowOverlayCommand } from './showOverlay'
 
 interface SetStackRootCommand {
   <P = {}>(layout: Layout<P> | Layout<P>[]): Promise<any>
@@ -19,22 +17,9 @@ interface PushCommand {
   <P = {}>(name: string, passProps?: P, options?: Options): Promise<any>
 }
 
-interface ShowModalCommand {
-  (layout: Layout): Promise<any>
-  <P = {}>(name: string, passProps?: P, options?: Options): Promise<any>
-}
-
-interface ShowOverlayCommand {
-  (layout: Layout): Promise<any>
-  <P = {}>(name: string, passProps?: P, options?: Options): Promise<any>
-}
-
 export type NavigationCommands = {
-  setRoot: SetRootCommand
   setStackRoot: SetStackRootCommand
   push: PushCommand
-  showModal: ShowModalCommand
-  showOverlay: ShowOverlayCommand
   mergeOptions: (options: Options) => void
   updateProps: (props: object) => void
   dismissModal: (mergeOptions?: Options) => Promise<any>
@@ -42,8 +27,35 @@ export type NavigationCommands = {
   popTo: (mergeOptions?: Options) => Promise<any>
   popToRoot: (mergeOptions?: Options) => Promise<any>
   dismissOverlay: () => Promise<any>
+
+  /**
+   * @deprecated Use setRoot import from 'react-native-navigation-hooks'
+   */
+  setRoot: SetRootCommand
+
+  /**
+   * @deprecated Use setRoot import from 'react-native-navigation-hooks'
+   */
+  showModal: ShowModalCommand
+
+  /**
+   * @deprecated Use showOverlay import from 'react-native-navigation-hooks'
+   */
+  showOverlay: ShowOverlayCommand
+
+  /**
+   * @deprecated Use Navigation.setDefaultOptions instead
+   */
   setDefaultOptions: (options: Options) => void
+
+  /**
+   * @deprecated Use Navigation.dismissAllModals instead
+   */
   dismissAllModals: (mergeOptions?: Options) => Promise<any>
+
+  /**
+   * @deprecated Use Navigation.getLaunchArgs instead
+   */
   getLaunchArgs: () => Promise<any>
 }
 
@@ -62,20 +74,6 @@ function createNavigationCommands(
    */
   componentId: string
 ): NavigationCommands {
-  function setRoot<P = {}>(nameOrLayout: string | Layout | LayoutRoot, passProps?: P, options?: Options) {
-    let layoutRoot
-
-    if (typeof nameOrLayout === 'string') {
-      layoutRoot = { root: createLayout<P>(nameOrLayout, passProps, options) } as LayoutRoot
-    } else if ((nameOrLayout as LayoutRoot).root) {
-      layoutRoot = nameOrLayout as LayoutRoot
-    } else {
-      layoutRoot = { root: nameOrLayout } as LayoutRoot
-    }
-
-    return Navigation.setRoot(layoutRoot)
-  }
-
   function setStackRoot<P = {}>(nameOrLayout: string | Layout<P> | Array<Layout<P>>, passProps?: P, options?: Options) {
     const layout = typeof nameOrLayout === 'string' ? createLayout<P>(nameOrLayout, passProps, options) : nameOrLayout
 
@@ -86,18 +84,6 @@ function createNavigationCommands(
     const layout = typeof nameOrLayout === 'string' ? createLayout<P>(nameOrLayout, passProps, options) : nameOrLayout
 
     return Navigation.push(componentId, layout)
-  }
-
-  function showModal<P = {}>(nameOrLayout: string | Layout<P>, passProps?: P, options?: Options) {
-    const layout = typeof nameOrLayout === 'string' ? createLayout<P>(nameOrLayout, passProps, options) : nameOrLayout
-
-    return Navigation.showModal(layout)
-  }
-
-  function showOverlay<P = {}>(nameOrLayout: string | Layout<P>, passProps?: P, options?: Options) {
-    const layout = typeof nameOrLayout === 'string' ? createLayout<P>(nameOrLayout, passProps, options) : nameOrLayout
-
-    return Navigation.showOverlay(layout)
   }
 
   function mergeOptions(options: Options) {
@@ -135,11 +121,8 @@ function createNavigationCommands(
   const { setDefaultOptions, dismissAllModals, getLaunchArgs } = Navigation
 
   return {
-    setRoot,
     setStackRoot,
     push,
-    showModal,
-    showOverlay,
     mergeOptions,
     updateProps,
     dismissModal,
@@ -147,6 +130,9 @@ function createNavigationCommands(
     popTo,
     popToRoot,
     dismissOverlay,
+    setRoot,
+    showModal,
+    showOverlay,
     setDefaultOptions,
     dismissAllModals,
     getLaunchArgs,
