+++ src/helpers/index.ts
@@ -1,3 +1,6 @@
 export { default as createLayout } from './createLayout'
 export { default as createStack } from './createStack'
 export { default as createNavigationCommands } from './createNavigationCommands'
+export { default as setRoot } from './setRoot'
+export { default as showModal } from './showModal'
+export { default as showOverlay } from './showOverlay'
